requirements.txt
sqlalchemy
psycopg2-binary
python-dotenv
streamlit
plotly
pandas
streamlit-mic-recorder
pydub
groq
edge-tts